AP-11836: Fix a race condition in the reader
The following race condition could happen if flow variables were used to
set the file location:

1. During the loading of the settings the AtomicBoolean
currentlyLoadingSettings is set to true.
2. As a last step of loading the settings the sheet list is updated in a
background thread.
3. While the background thread was running the flow variables were
injected and the file location changed. This didn't cause any update
because currentlyLoadingSettings is still true.
4. Once the background thread is finished the variable
currentlyLoadingSettings is set to false by the change listener
5. The background thread sets the selected sheet to null (because the file
which was seleceted when the thread started doesn't exist). Note that
updateFileTable sets the status to fileNotFound() if the sheet is null.

--> This leaves the dialog in the invalid state

This commit solves this race condition by setting the variable
currentlyLoadingSettings to false before the background thread finishes.

This can cause multiple background threads to be running and to prevent
an outdated update if one background thread overtakes another one a
variable is added which saves an index of the most current thread.