AP-17559: Fixed formula evaluation Excel Writer
Fixed a problem with the formula evaluation in case of XLSX files and
the Excel Writer by switching from SXSSFWorkbooks to XSSFWorkbooks
(which is more memory intensive) in case formula evaluation is checked.
SXSSFWorkbooks does not support formula evaluation.

AP-17559 (Excel Writer fails when writing to XLSX with Evaluate Formulas option enabled)