AP-20714: Fix local caching logic
The problem was that the method used was `orElse` which eagerly evaluated its argument,
scheduling the caching... (this is why the debug log message would always appear)
The fix is to use `orElseGet` instead.

The rest of this commit is just to refactor the temp file copy logic into its own method (no functional change).

AP-20714 (Excel Reader supporting reading "large" files (300+ MB))